requirements.txt
GPUtil
einops
matplotlib
ml-collections
monai
pytorch
pytorch-cuda=11.7
scikit-image
scikit-learn
tensorboard
timm
torchaudio
torchvision
tqdm